requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
boto3==1.28.71
python-dotenv==1.0.0
mysql-connector-python==8.1.0
